+++ RELEASE_NOTES.md
@@ -1,4 +1,9 @@
 Improvement release
 
 **Bugfixes**
-- Security update phpunit
+- Fix Absence Type statistics
+- Fix Presence Type statistics
+- Fix Remainder statistics
+
+**Improvements**
+- Mobile devices (different screen widths) are now automatically supported